refacotr : image 변경 모달 뷰로직 분리
