+++ src/1_Page/Profile/ui/Header/ui/ModifyImageModal/index.js
@@ -5,56 +5,12 @@ import Modal from "../../../../../../2_Widget/Modal";
 import ConfirmModal from "../../../../../../2_Widget/ConfirmModal";
 
 import useFileReader from "./model/useFileReader";
-import useConfirmModal from "../../../../model/useConfirmModal";
-import useModifyImage from "./api/useModifyImage";
+import useImageModal from "./model/useImageModal";
 
 const ModifyImageModal = (props) => {
   const { onClose } = props;
   const { image } = props;
 
-  const [message, setMessage] = useState("");
-
-  const { confirmModal, handleConfirmModalOpen, handleConfirmModalClose } =
-    useConfirmModal();
-
-  const { modify, loading, error } = useModifyImage();
-
-  const closeRef = useRef(null);
-  const imageRef = useRef(null);
-
-  useEffect(() => {
-    imageRef.current = image;
-  }, [image]);
-
-  const handleImageConfirmModalOpen = (handleClose) => {
-    handleConfirmModalOpen();
-    closeRef.current = handleClose;
-  };
-
-  const handleModifyClick = async (handleClose) => {
-    if (errorMessage) {
-      setMessage(errorMessage);
-      handleConfirmModalOpen();
-    }
-    if (imageRef.current == imagePreview) {
-      setMessage("사진을 변경하세요");
-      handleConfirmModalOpen();
-      return;
-    }
-    const result = await modify(imagePreview);
-    if (result) {
-      setMessage("변경되었습니다");
-      handleImageConfirmModalOpen(handleClose);
-      return;
-    }
-    setMessage(result);
-    handleConfirmModalOpen();
-  };
-
-  const handleImageConfirmModalDone = () => {
-    handleConfirmModalClose();
-    if (closeRef.current) closeRef.current();
-  };
   const {
     fileInputRef,
     imagePreview,
@@ -62,6 +18,13 @@ const ModifyImageModal = (props) => {
     handleProfileImageClick,
     handleFileChange,
   } = useFileReader(image);
+
+  const {
+    message,
+    confirmModal,
+    handleModifyClick,
+    handleImageConfirmModalDone,
+  } = useImageModal(image, errorMessage, imagePreview);
 
   return (
     <>
